Moved JobExecutionThinResource to resources package (#2820)
resolves #2801
